requirements.txt
streamlit
requests
python-dotenv
supabase
langwatch
streamlit-authenticator
pandas
